backend/requirements.txt
# FastAPI Backend Dependencies

# Web Framework
fastapi

# ASGI Server
uvicorn[standard]

# ORM & Database
sqlmodel
alembic
asyncpg

# Authentication & Security
python-jose[cryptography]
passlib[bcrypt]
bcrypt

# Form Data Handling
python-multipart

# Environment Variables
python-dotenv
